# Chart Loader Optimization